+++ lib/lightning/invocation/run.ex
@@ -7,12 +7,16 @@ defmodule Lightning.Invocation.Run do
   """
   use Ecto.Schema
   import Ecto.Changeset
-  alias Lightning.Invocation.Event
+  alias Lightning.Invocation.{Event, Dataclip}
+  alias Lightning.Projects.Project
+  alias Lightning.Jobs.Job
 
   @type t :: %__MODULE__{
           __meta__: Ecto.Schema.Metadata.t(),
           id: Ecto.UUID.t() | nil,
-          event: Event.t() | Ecto.Association.NotLoaded.t() | nil
+          event: Event.t() | Ecto.Association.NotLoaded.t() | nil,
+          project: Project.t() | Ecto.Association.NotLoaded.t() | nil,
+          job: Job.t() | Ecto.Association.NotLoaded.t() | nil
         }
 
   @primary_key {:id, :binary_id, autogenerate: true}
@@ -23,11 +27,15 @@ defmodule Lightning.Invocation.Run do
     field :log, {:array, :string}
     field :started_at, :utc_datetime_usec
     belongs_to :event, Event
-    has_one :job, through: [:event, :job]
-    has_one :project, through: [:event, :project]
-    has_one :source_dataclip, through: [:event, :dataclip]
 
-    has_one :result_dataclip, through: [:event, :result_dataclip]
+    belongs_to :project, Project
+    belongs_to :job, Job
+
+    belongs_to :input_dataclip, Dataclip
+    belongs_to :output_dataclip, Dataclip
+
+    # has_one :source_dataclip, through: [:event, :dataclip]
+    # has_one :result_dataclip, through: [:event, :result_dataclip]
 
     timestamps(usec: true)
   end
@@ -35,8 +43,22 @@ defmodule Lightning.Invocation.Run do
   @doc false
   def changeset(run, attrs) do
     run
-    |> cast(attrs, [:log, :exit_code, :started_at, :finished_at, :event_id])
+    |> cast(attrs, [
+      :log,
+      :exit_code,
+      :started_at,
+      :finished_at,
+      :event_id,
+      :project_id,
+      :job_id,
+      :input_dataclip_id,
+      :output_dataclip_id
+    ])
+    |> cast_assoc(:output_dataclip, with: &Dataclip.changeset/2, required: false)
     |> foreign_key_constraint(:event_id)
-    |> validate_required([:event_id])
+    |> foreign_key_constraint(:input_dataclip_id)
+    |> foreign_key_constraint(:output_dataclip_id)
+    |> foreign_key_constraint(:job_id)
+    |> validate_required([:event_id, :project_id, :job_id, :input_dataclip_id])
   end
 end