+++ test/lightning/invocation_test.exs
@@ -5,6 +5,7 @@ defmodule Lightning.InvocationTest do
   alias Lightning.Repo
   import Lightning.InvocationFixtures
   import Lightning.ProjectsFixtures
+  import Lightning.JobsFixtures
 
   describe "invocation" do
     import Lightning.JobsFixtures
@@ -72,7 +73,7 @@ defmodule Lightning.InvocationTest do
       assert Invocation.get_dataclip(dataclip.id) == dataclip
       assert Invocation.get_dataclip(Ecto.UUID.generate()) == nil
 
-      run = run_fixture(event_id: event.id)
+      run = run_fixture(event_id: event.id, input_dataclip_id: dataclip.id)
 
       assert Invocation.get_dataclip(run) == dataclip
     end
@@ -187,6 +188,7 @@ defmodule Lightning.InvocationTest do
     alias Lightning.Invocation.Run
 
     import Lightning.InvocationFixtures
+    import Lightning.ProjectsFixtures
 
     @invalid_attrs %{event_id: nil}
     @valid_attrs %{
@@ -206,11 +208,13 @@ defmodule Lightning.InvocationTest do
       event = event_fixture(project_id: project.id)
 
       first_run =
-        run_fixture(event_id: event.id)
+        run_fixture(event_id: event.id, project_id: project.id)
         |> shift_inserted_at!(days: -1)
         |> Repo.preload(:job)
 
-      second_run = run_fixture(event_id: event.id) |> Repo.preload(:job)
+      second_run =
+        run_fixture(event_id: event.id, project_id: project.id)
+        |> Repo.preload(:job)
 
       assert Invocation.list_runs_for_project(project).entries == [
                second_run,
@@ -230,11 +234,20 @@ defmodule Lightning.InvocationTest do
     end
 
     test "create_run/1 with valid data creates a run" do
-      event = event_fixture()
+      project = project_fixture()
+      dataclip = dataclip_fixture(project_id: project.id)
+      job = job_fixture(project_id: project.id)
+
+      event = event_fixture(project_id: project.id, dataclip_id: dataclip.id)
 
       assert {:ok, %Run{} = run} =
                Invocation.create_run(
-                 Map.merge(@valid_attrs, %{event_id: event.id})
+                 Map.merge(@valid_attrs, %{
+                   event_id: event.id,
+                   project_id: project.id,
+                   job_id: job.id,
+                   input_dataclip_id: dataclip.id
+                 })
                )
 
       assert run.exit_code == 42