+++ test/support/fixtures/invocation_fixtures.ex
@@ -54,9 +54,21 @@ defmodule Lightning.InvocationFixtures do
   def run_fixture(attrs \\ []) when is_list(attrs) do
     {event_attrs, attrs} = Keyword.pop(attrs, :event_attrs, [])
 
+    attrs =
+      attrs
+      |> Keyword.put_new_lazy(:project_id, fn ->
+        Lightning.ProjectsFixtures.project_fixture().id
+      end)
+
     {:ok, run} =
       attrs
+      |> Keyword.put_new_lazy(:job_id, fn ->
+        Lightning.JobsFixtures.job_fixture(project_id: attrs[:project_id]).id
+      end)
       |> Keyword.put_new_lazy(:event_id, fn -> event_fixture(event_attrs).id end)
+      |> Keyword.put_new_lazy(:input_dataclip_id, fn ->
+        dataclip_fixture(project_id: attrs[:project_id]).id
+      end)
       |> Enum.into(%{
         exit_code: nil,
         finished_at: nil,